Size Type changed in Space

+++ db/migrate/20190704104136_add_size_to_spaces.rb
@@ -1,5 +1,5 @@
 class AddSizeToSpaces < ActiveRecord::Migration[5.2]
   def change
-    add_column :spaces, :size, :integer
+    add_column :spaces, :size, :float
   end
 end
